Rename call view providers in parallel handler
